Replace Contributor delete prompt with PureComponent (and tweak modal-background behaviour)
